extract request span options into separate file and dedupe isStreamingHandler
