+++ packages/aws-serverless/src/sdk.ts
@@ -139,7 +139,7 @@ export const AWS_HANDLER_HIGHWATERMARK_SYMBOL = Symbol.for('aws.lambda.runtime.h
 export const AWS_HANDLER_STREAMING_SYMBOL = Symbol.for('aws.lambda.runtime.handler.streaming');
 export const AWS_HANDLER_STREAMING_RESPONSE = 'response';
 
-function isStreamingHandler(handler: Handler | StreamifyHandler): handler is StreamifyHandler {
+export function isStreamingHandler(handler: Handler | StreamifyHandler): handler is StreamifyHandler {
   return (
     (handler as unknown as Record<symbol, unknown>)[AWS_HANDLER_STREAMING_SYMBOL] === AWS_HANDLER_STREAMING_RESPONSE
   );
